+++ src/placementEngine.test.js
@@ -1,53 +1,177 @@
 import placementEngine from './placementEngine'
-import stateMachine from './stateMachine'
-import { get } from './userState'
+import { get, set } from './userState'
 
-jest.mock('./stateMachine', () => jest.fn())
-jest.mock('./userState', () => ({ get: jest.fn() }))
+// Mock the placePromos function to return the promos unchanged.
+jest.mock('./placePromos', () => jest.fn(promos => promos))
+
+jest.mock('./userState', () => ({
+  get: jest.fn(),
+  set: jest.fn()
+}))
+
+// Mock the timestamp function.
+jest.mock('./utils', () => ({ timestamp: jest.fn(() => '123') }))
 
 describe('placementEngine', () => {
-  const user = { visits: 1 }
-  const promo = { promoid: 1 }
-  const promos = [promo]
+  const promos = [
+    { promoId: 1, groupId: 1, campaignId: 1 },
+    { promoId: 2, groupId: 1, campaignId: 1 },
+    { promoId: 3, campaignId: 1 }
+  ]
 
-  get.mockReturnValue(user)
+  const state = {
+    blocked: {},
+    impressions: {},
+    engagements: {},
+    visits: 0
+  }
 
-  const innerMock = jest.fn((state, event, emit) => {
-    emit.next({ promos })
-    return state
+  get.mockReturnValue(state)
+
+  it('returns a promise that resolves the placed promos', () => {
+    const result = placementEngine(promos, window)
+    return expect(result).resolves.toHaveProperty('promos', promos)
   })
 
-  // The stateMachine function is curried, so we need to return the inner mocked
-  // function.
-  stateMachine.mockImplementation(() => innerMock)
-
-  it('emits an initial init event to the state machine', done => {
-    placementEngine([], window)
-      .subscribe(state => {
-        expect(state.promos).toEqual(promos)
-        expect(stateMachine).toHaveBeenLastCalledWith([], window)
-        expect(innerMock).toHaveBeenLastCalledWith({ user }, { type: 'visit' }, expect.anything())
-        done()
+  it('increments the number of visits', () => {
+    placementEngine(promos, window)
+    expect(set).toHaveBeenLastCalledWith(
+      window.localStorage,
+      expect.objectContaining({
+        visits: 1
       })
+    )
   })
 
-  it('handles the onClick/onClose callbacks', done => {
-    let onClick, onClose
-
-    placementEngine([], window)
-      .subscribe(state => {
-        onClick = state.onClick
-        onClose = state.onClose
+  describe('onClick', () => {
+    it('updates the campaign engagements', () => {
+      return placementEngine(promos, window).then(({ onClick }) => {
+        onClick(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            engagements: expect.objectContaining({
+              campaigns: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
       })
+    })
 
-    setTimeout(() => {
-      onClick(promo)
-      expect(innerMock).toHaveBeenLastCalledWith({ user }, { type: 'click', promo }, expect.anything())
+    it('updates the group engagements', () => {
+      return placementEngine(promos, window).then(({ onClick }) => {
+        onClick(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            engagements: expect.objectContaining({
+              groups: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
+      })
+    })
 
-      onClose(promo)
-      expect(innerMock).toHaveBeenLastCalledWith({ user }, { type: 'close', promo }, expect.anything())
+    it('updates the promo engagements', () => {
+      return placementEngine(promos, window).then(({ onClick }) => {
+        onClick(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            engagements: expect.objectContaining({
+              promos: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
+      })
+    })
+  })
 
-      done()
-    }, 0)
+  describe('onClose', () => {
+    it('updates the blocked campaigns', () => {
+      return placementEngine(promos, window).then(({ onClose }) => {
+        onClose(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            blocked: expect.objectContaining({
+              campaigns: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
+      })
+    })
+
+    it('updates the blocked groups', () => {
+      return placementEngine(promos, window).then(({ onClose }) => {
+        onClose(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            blocked: expect.objectContaining({
+              groups: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
+      })
+    })
+
+    it('updates the blocked promos', () => {
+      return placementEngine(promos, window).then(({ onClose }) => {
+        onClose(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            blocked: expect.objectContaining({
+              promos: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
+      })
+    })
+  })
+
+  describe('onView', () => {
+    it('updates the campaign impressions', () => {
+      return placementEngine(promos, window).then(({ onView }) => {
+        onView(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            impressions: expect.objectContaining({
+              campaigns: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
+      })
+    })
+
+    it('updates the group impressions', () => {
+      return placementEngine(promos, window).then(({ onView }) => {
+        onView(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            impressions: expect.objectContaining({
+              groups: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
+      })
+    })
+
+    it('updates the promo impressions', () => {
+      return placementEngine(promos, window).then(({ onView }) => {
+        onView(promos[0])
+        expect(set).toHaveBeenLastCalledWith(
+          window.localStorage,
+          expect.objectContaining({
+            impressions: expect.objectContaining({
+              promos: { 1: { count: 1, timestamp: '123' } }
+            })
+          })
+        )
+      })
+    })
   })
 })