+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Return user from userState `set` function
+* Refactor the placement engine to use promises
 
 ## 3.0.0 (2019-04-30)